requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
requests==2.31.0
httpx==0.25.2
asyncpg==0.29.0
databases[postgresql]==0.8.0
aiofiles==23.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
